Revert directional FOV and change attacks to square areas

+++ Roguelike/Items/ItemParameter.cs
@@ -11,8 +11,8 @@ namespace Roguelike.Items
 
         public int AttackSpeed { get; set; } = Constants.FULL_TURN;
         public int Damage { get; set; } = Constants.DEFAULT_DAMAGE;
-        public double MeleeRange { get; set; } = Constants.DEFAULT_MELEE_RANGE;
-        public double ThrowRange { get; set; } = Constants.DEFAULT_THROW_RANGE;
+        public int MeleeRange { get; set; } = Constants.DEFAULT_MELEE_RANGE;
+        public int ThrowRange { get; set; } = Constants.DEFAULT_THROW_RANGE;
 
         public ItemParameter(string name, MaterialType material)
         {
